move pool creation fee test to swaprouter
